Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
pymongo==4.6.0
Flask-PyMongo==2.3.0
Flask==3.0.0
gunicorn==21.2.0
jaeger-client==4.8.0
prometheus-client==0.18.0
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation==0.42b0
opentelemetry-sdk==1.21.0
prometheus-flask-exporter==0.23.0
pylint==3.0.2
requests==2.31.0
Flask-Cors==4.0.0
opentelemetry-exporter-prometheus==1.12.0rc1